% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diff_heatmap_plot.R
\name{create_diff_heatmap_plot}
\alias{create_diff_heatmap_plot}
\title{Create a Differential Heatmap Plot}
\usage{
create_diff_heatmap_plot(
  data,
  var_1_name,
  var_2_name,
  start_year,
  end_year,
  x_axis_breaks,
  title,
  frequency,
  change_space,
  flip_colors = FALSE,
  override_subtitle = NULL,
  include_cell_numbers = FALSE,
  scaling_power = NULL,
  num_decimals = NULL
)
}
\arguments{
\item{data}{A data frame containing the dataset for plotting. Must include a 'date' column.}

\item{var_1_name}{Name of the first variable in 'data' to be used in the differential calculation.}

\item{var_2_name}{Name of the second variable in 'data' to be used in the differential calculation.}

\item{start_year}{The starting year for the plot.}

\item{end_year}{The ending year for the plot.}

\item{x_axis_breaks}{Breaks for the x-axis, usually years.}

\item{title}{Title of the heatmap.}

\item{frequency}{The frequency of data points ('weekly' or 'monthly').}

\item{change_space}{Determines the type of data transformation ('level' or 'yoy' for year-over-year).}

\item{flip_colors}{A boolean indicating whether to flip the default color scheme.}

\item{override_subtitle}{An optional subtitle for the heatmap, overriding the default.}

\item{include_cell_numbers}{A boolean indicating whether to include cell numbers in the heatmap.}

\item{scaling_power}{The power to which the cell numbers are scaled (relevant only if include_cell_numbers is TRUE).}

\item{num_decimals}{The number of decimal places to include in cell number labels (if applicable).}
}
\value{
A ggplot object representing the differential heatmap.
}
\description{
This function generates a heatmap plot that visualizes the difference between two variables over time. It integrates recession data and offers various customization options.
}
\examples{
# Assuming 'data' is your dataset with required columns
diff_heatmap_plot <- create_diff_heatmap_plot(data, "var1", "var2", 2010, 2020, 1, "Differential Heatmap Title",
                                              "monthly", "level", FALSE, NULL, FALSE, NULL, 2)
print(diff_heatmap_plot)
}
