% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap_plot.R
\name{create_heatmap_plot}
\alias{create_heatmap_plot}
\title{Create a Heatmap Plot}
\usage{
create_heatmap_plot(
  data,
  var_name,
  start_year,
  end_year,
  x_axis_breaks,
  title,
  frequency,
  change_space,
  flip_colors = FALSE,
  override_subtitle = NULL,
  include_cell_numbers = FALSE,
  scaling_power = NULL,
  num_decimals = NULL
)
}
\arguments{
\item{data}{A data frame that contains the data to be plotted. Must include a 'date' column.}

\item{var_name}{Name of the variable in 'data' to plot.}

\item{start_year}{The starting year for the plot.}

\item{end_year}{The ending year for the plot.}

\item{x_axis_breaks}{Breaks for the x-axis, usually years.}

\item{title}{Title of the heatmap.}

\item{frequency}{The frequency of data points ('weekly' or 'monthly').}

\item{change_space}{Determines the type of data transformation ('level' or 'yoy' for year-over-year).}

\item{flip_colors}{A boolean indicating whether to flip the default color scheme.}

\item{override_subtitle}{An optional subtitle for the heatmap, overriding the default.}

\item{include_cell_numbers}{A boolean indicating whether to include cell numbers in the heatmap.}

\item{scaling_power}{The power to which the cell numbers are scaled (relevant only if include_cell_numbers is TRUE).}

\item{num_decimals}{The number of decimal places to include in cell number labels (if applicable).}
}
\value{
A ggplot object representing the heatmap.
}
\description{
This function generates a heatmap plot based on the specified parameters. It integrates recession data and supports various customization options for the heatmap.
}
\examples{
# Assuming 'data' is your dataset with required columns
heatmap_plot <- create_heatmap_plot(data, "initial_claims_nsa", 2010, 2020, 1, "Initial Claims",
                                   "monthly", "level", FALSE, NULL, FALSE, NULL, 2)
print(heatmap_plot)
}
