% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_returns_plot.R
\name{create_returns_plot}
\alias{create_returns_plot}
\title{Create a Returns Plot}
\usage{
create_returns_plot(
  price_data,
  start_date,
  end_date = Sys.Date(),
  variables,
  combo_bets = NULL,
  add_risk_parity = FALSE
)
}
\arguments{
\item{price_data}{A data frame with a column of dates and columns representing price data for various financial instruments.}

\item{start_date}{The start date for calculating returns, formatted as 'YYYY-MM-DD'.}

\item{end_date}{The end date for calculating returns, formatted as 'YYYY-MM-DD'. Default is the current system date.}

\item{variables}{A character vector of column names in 'price_data' representing the financial instruments for which returns are to be calculated.}

\item{combo_bets}{An optional list of lists, each containing information about a combination bet, including variables and their respective sizes.}

\item{add_risk_parity}{A logical flag indicating whether to add a column calculating risk parity based returns. Default is FALSE.}
}
\value{
A ggplot object representing the returns plot.
}
\description{
This function generates a plot visualizing the returns of specified financial instruments over a given time period. It also supports the visualization of combo bets and an optional risk parity calculation.
}
\examples{
# Assuming 'my_price_data' is your dataset with required columns
returns_plot <- create_returns_plot(my_price_data, "2020-01-01", "2020-12-31", c("AAPL", "MSFT"), NULL, FALSE)
print(returns_plot)
}
