% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_returns_table.R
\name{create_returns_table}
\alias{create_returns_table}
\title{Create Returns Table for Use in Returns Plot}
\usage{
create_returns_table(
  price_data,
  start_date,
  end_date = Sys.Date(),
  variables,
  combo_bets = NULL,
  add_risk_parity = FALSE
)
}
\arguments{
\item{price_data}{A data frame with a column named 'date' and other columns representing price data for various financial instruments.}

\item{start_date}{The start date for calculating returns, formatted as 'YYYY-MM-DD'.}

\item{end_date}{The end date for calculating returns, formatted as 'YYYY-MM-DD'. Default is the current system date.}

\item{variables}{A character vector of column names in 'price_data' for which returns are to be calculated.}

\item{combo_bets}{An optional list of lists, each containing information about a combination bet, including variables and their respective sizes.}

\item{add_risk_parity}{A logical flag indicating whether to add a column calculating risk parity based returns. Default is FALSE.}
}
\value{
A data frame where each row corresponds to a date and each column to the returns of a specified variable, combination bet, or risk parity calculation.
}
\description{
This function calculates returns for given financial variables over a specified time range, handling individual variables, combinations of variables ("combo bets"), and an optional risk parity calculation.
}
\examples{
# Assuming 'my_price_data' is your dataset with required columns
start_date <- "2020-01-01"
end_date <- "2020-12-31"
variables <- c("AAPL", "MSFT")
combo_bets <- list(
  list(var_1 = "AAPL", var_2 = "MSFT", size_1 = 0.5, size_2 = 0.5)
)
returns_table <- create_returns_table(my_price_data, start_date, end_date, variables, combo_bets)
}
