% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter_plot.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Create a Scatter Plot}
\usage{
create_scatter_plot(
  data,
  x_var,
  y_var,
  x_label,
  y_label,
  title = NULL,
  highlight_dates = NULL,
  highlight_latest = FALSE,
  include_lof = NULL,
  log_x = FALSE,
  log_y = FALSE,
  start_date = NULL,
  end_date = NULL,
  x_yoy = FALSE,
  y_yoy = FALSE,
  date_ranges = NULL
)
}
\arguments{
\item{data}{A data frame containing the dataset for plotting.}

\item{x_var}{Name of the variable to be plotted on the x-axis.}

\item{y_var}{Name of the variable to be plotted on the y-axis.}

\item{x_label}{Label for the x-axis.}

\item{y_label}{Label for the y-axis.}

\item{title}{Optional title for the plot.}

\item{highlight_dates}{Dates to highlight on the plot.}

\item{highlight_latest}{A boolean indicating whether to highlight the latest data point.}

\item{include_lof}{Method for line of best fit ('lm', 'loess', or NULL).}

\item{log_x}{A boolean indicating whether to apply a logarithmic transformation to the x-axis.}

\item{log_y}{A boolean indicating whether to apply a logarithmic transformation to the y-axis.}

\item{start_date}{The starting date for the plot (optional).}

\item{end_date}{The ending date for the plot (optional).}

\item{x_yoy}{A boolean indicating whether to calculate year-over-year change for the x variable.}

\item{y_yoy}{A boolean indicating whether to calculate year-over-year change for the y variable.}

\item{date_ranges}{A list of date ranges to highlight on the plot.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function generates a scatter plot for the relationship between two variables over time. It includes options for highlighting specific dates, adding a line of best fit, and log transformations.
}
\examples{
# Assuming 'data' is your dataset with required columns
scatter_plot <- create_scatter_plot(data, "x_variable", "y_variable", "X Axis Label", "Y Axis Label",
                                   "Scatter Plot Title", c("2020-01-01", "2021-01-01"), FALSE, NULL, FALSE, FALSE)
print(scatter_plot)
}
