% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_two_axis_line_plot.R
\name{create_two_axis_line_plot}
\alias{create_two_axis_line_plot}
\title{Create a Two-Axis Line Plot}
\usage{
create_two_axis_line_plot(
  data,
  variables,
  var_labels,
  start_date,
  end_date,
  plot_title,
  x_axis_breaks,
  var_1_type,
  var_2_type
)
}
\arguments{
\item{data}{A data frame containing the dataset for plotting.}

\item{variables}{A vector of two variable names in 'data' to be plotted.}

\item{var_labels}{A vector of two labels corresponding to the variables.}

\item{start_date}{The starting date for the plot, formatted as 'YYYY-MM-DD'.}

\item{end_date}{The ending date for the plot, formatted as 'YYYY-MM-DD'.}

\item{plot_title}{Title of the plot.}

\item{x_axis_breaks}{Breaks for the x-axis, usually years.}

\item{var_1_type}{Type of transformation for the first variable ('level' or 'yoy' for year-over-year).}

\item{var_2_type}{Type of transformation for the second variable ('level' or 'yoy' for year-over-year).}
}
\value{
A ggplot object representing the two-axis line plot.
}
\description{
This function generates a line plot for two variables, each with its own y-axis, over a given time period. The plot includes recession bars and supports different types of data transformations.
}
\examples{
# Assuming 'data' is your dataset with required columns
two_axis_plot <- create_two_axis_line_plot(data, c("var1", "var2"), c("Variable 1", "Variable 2"),
                                           "2020-01-01", "2020-12-31", "Two-Axis Line Plot", 1, "level", "yoy")
print(two_axis_plot)
}
