% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cumulative_table.R
\name{get_cumulative_table}
\alias{get_cumulative_table}
\title{Calculate Cumulative Percentage Change for Variables Over Specified Intervals}
\usage{
get_cumulative_table(
  df,
  var_names,
  interval_short = NULL,
  interval_manual = NULL
)
}
\arguments{
\item{df}{A dataframe containing a date column and other numeric columns for calculation.}

\item{var_names}{A vector of strings specifying the column names in the dataframe for which
the cumulative percentage change is to be calculated.}

\item{interval_short}{A string specifying the type of interval for the calculation.
It can be "decades" or "years". If this parameter is used,
`interval_manual` must be NULL.}

\item{interval_manual}{A list of numeric vectors, each containing two elements
representing the start and end years of the intervals.
If this parameter is used, `interval_short` must be NULL.}
}
\value{
A tibble with rows representing the intervals and columns showing the
        cumulative percentage change for the specified variables.
}
\description{
This function computes the cumulative percentage change for given variables
in a dataframe over specified time intervals.
}
\examples{
# Assuming df is a dataframe with a date column and numeric columns "var1" and "var2"
df <- data.frame(
  date = seq(as.Date("2000-01-01"), by = "year", length.out = 21),
  var1 = runif(21, 50, 100),
  var2 = runif(21, 20, 80)
)
results_decades <- get_cumulative_table(df, c("var1", "var2"), interval_short = "decades")
results_manual <- get_cumulative_table(df, c("var1", "var2"), interval_manual = list(c(2000, 2009), c(2010, 2019)))
}
