% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap_plot.R
\name{create_heatmap_plot}
\alias{create_heatmap_plot}
\title{Create a Heatmap Plot}
\usage{
create_heatmap_plot(
  data,
  var_name,
  start_year,
  end_year,
  x_axis_breaks,
  title,
  frequency,
  change_space,
  flip_colors = FALSE,
  override_subtitle = NULL,
  include_cell_numbers = FALSE,
  scaling_power = NULL,
  num_decimals = NULL,
  color_1 = "indianred",
  color_2 = "dodgerblue",
  mute_color_1 = FALSE,
  mute_color_2 = TRUE,
  legend_name = "Capped Level",
  x_axis_label = NULL,
  y_axis_label = NULL,
  text_size = 3,
  text_bold = FALSE
)
}
\arguments{
\item{data}{A data frame that contains the data to be plotted. Must include a 'date' column.}

\item{var_name}{Name of the variable in 'data' to plot.}

\item{start_year}{The starting year for the plot.}

\item{end_year}{The ending year for the plot.}

\item{x_axis_breaks}{Breaks for the x-axis, usually years.}

\item{title}{Title of the heatmap.}

\item{frequency}{The frequency of data points ('weekly' or 'monthly').}

\item{change_space}{Determines the type of data transformation ('level' or 'yoy' for year-over-year).}

\item{flip_colors}{A boolean indicating whether to flip the color scheme.}

\item{override_subtitle}{An optional subtitle for the heatmap, overriding the default.}

\item{include_cell_numbers}{A boolean indicating whether to include cell numbers in the heatmap.}

\item{scaling_power}{The power to which the cell numbers are scaled (relevant only if include_cell_numbers is TRUE).}

\item{num_decimals}{The number of decimal places to include in cell number labels (if applicable).}

\item{color_1}{The primary color used in the heatmap. This can be any valid color name or hexadecimal color code. The default is "indianred". This color is used for the low end of the gradient scale unless `flip_colors` is TRUE.}

\item{color_2}{The secondary color used in the heatmap. This can be any valid color name or hexadecimal color code. The default is "dodgerblue". This color is used for the high end of the gradient scale unless `flip_colors` is TRUE.}

\item{mute_color_1}{A boolean indicating whether to mute `color_1`. If TRUE, `color_1` will be muted, which typically makes it less bright and more pastel-like. Default is FALSE.}

\item{mute_color_2}{A boolean indicating whether to mute `color_2`. If TRUE, `color_2` will be muted, which typically makes it less bright and more pastel-like. Default is TRUE.}

\item{legend_name}{Title for the legend, defaulted as "Capped Level"}

\item{x_axis_label}{X axis label, no label default}

\item{y_axis_label}{T axis label, no label default}

\item{text_size}{Size of text in cells}

\item{text_bold}{Whether or not you want the text in the cells to be bold}
}
\value{
A ggplot object representing the heatmap.
}
\description{
This function generates a heatmap plot based on the specified parameters. It integrates recession data and supports various customization options for the heatmap.
}
\examples{
create_heatmap_plot(data = weekly_data, var_name = "initial_claims_nsa", start_year = 2000,
                    end_year = 2023, x_axis_breaks = 5, title = "Initial Claims NSA",
                    frequency = "weekly", change_space = "yoy", num_decimals = 1,
                    include_cell_numbers = TRUE, scaling_power = NULL, flip_colors = TRUE)

create_heatmap_plot(data = monthly_data, var_name = "initial_claims_nsa", start_year = 2000,
                    end_year = 2023, x_axis_breaks = 5, title = "Initial Claims NSA",
                    frequency = "monthly", change_space = "level", num_decimals = 1,
                    include_cell_numbers = TRUE, scaling_power = 5, flip_colors = TRUE)
}
