% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_line_plot.R
\name{create_line_plot}
\alias{create_line_plot}
\title{Create a Line Plot}
\usage{
create_line_plot(
  data,
  var_name,
  start_year,
  end_year,
  x_axis_breaks,
  x_axis_title,
  y_axis_title,
  title,
  plot_type,
  y_axis_breaks = NULL,
  include_smooth = TRUE,
  y_axis_lower_bound = NULL,
  y_axis_upper_bound = NULL
)
}
\arguments{
\item{data}{A data frame containing the dataset for plotting. Must include a 'date' column.}

\item{var_name}{Name of the variable in 'data' to be plotted.}

\item{start_year}{The starting year for the plot.}

\item{end_year}{The ending year for the plot.}

\item{x_axis_breaks}{Breaks for the x-axis, usually years.}

\item{x_axis_title}{Title for the x-axis.}

\item{y_axis_title}{Title for the y-axis.}

\item{title}{Title of the plot.}

\item{plot_type}{Type of the plot ('level' or 'yoy' for year-over-year).}

\item{y_axis_breaks}{Break points for the y-axis (optional).}

\item{include_smooth}{A boolean indicating whether to include a Loess smoothed line.}

\item{y_axis_lower_bound}{Lower bound for the y-axis (optional).}

\item{y_axis_upper_bound}{Upper bound for the y-axis (optional).}
}
\value{
A ggplot object representing the line plot.
}
\description{
This function generates a line plot for a given variable over time, with options for different types of plots and additional smoothing.
}
\examples{
# Assuming 'data' is your dataset with required columns
line_plot <- create_line_plot(data, "var1", 2010, 2020, 1, "X Axis Title", "Y Axis Title",
                             "Line Plot Title", "level", NULL, TRUE, NULL, NULL)
print(line_plot)
}
