% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multi_line_plot.R
\name{create_multi_line_plot}
\alias{create_multi_line_plot}
\title{Create a Multi-Line Plot}
\usage{
create_multi_line_plot(
  data,
  var_name_list,
  var_label_list,
  start_date,
  end_date,
  x_axis_breaks,
  use_month_intervals = FALSE,
  x_axis_title,
  y_axis_title,
  title,
  var_changes = NULL,
  y_axis_breaks = NULL,
  include_smooth = FALSE,
  lof_span = 0.1,
  subtitle = NULL,
  y_axis_lower_bound = NULL,
  y_axis_upper_bound = NULL,
  recession_bars = FALSE,
  diagonal_x_labels = FALSE
)
}
\arguments{
\item{data}{A data frame containing the dataset for plotting.
Must include a 'date' column.}

\item{var_name_list}{A vector of variable names in 'data' to be plotted.}

\item{var_label_list}{A vector of labels corresponding to the variables for the plot legend.}

\item{start_date}{The starting date for the plot.}

\item{end_date}{The ending date for the plot.}

\item{x_axis_breaks}{Breaks for the x-axis, can be years or months depending on 'use_month_intervals'.}

\item{use_month_intervals}{A boolean to determine if the x-axis should use month intervals.}

\item{x_axis_title}{Title for the x-axis.}

\item{y_axis_title}{Title for the y-axis.}

\item{title}{Title of the plot.}

\item{var_changes}{A list of transformations to apply to each variable.
Acceptable values are 'mom' for month-over-month, 'qoq' for quarter-over-quarter,
'yoy' for year-over-year, or 'level' for no change.
The list must be the same length as 'var_name_list'.}

\item{y_axis_breaks}{Break points for the y-axis (optional).}

\item{include_smooth}{A boolean indicating whether to include a Loess smoothed line.
Applicable only when no transformation is applied (i.e., var_changes is NULL or contains all NULLs).}

\item{lof_span}{Span for the loess smoothing (optional).}

\item{subtitle}{Subtitle for the plot (optional).}

\item{y_axis_lower_bound}{Lower bound for the y-axis (optional).}

\item{y_axis_upper_bound}{Upper bound for the y-axis (optional).}

\item{recession_bars}{A boolean indicating whether to include recession bars in the plot.}

\item{diagonal_x_labels}{A boolean indicating whether x-axis labels should be diagonal.}
}
\value{
A ggplot object representing the multi-line plot.
}
\description{
This function generates a multi-line plot for a list of variables over time,
with options for different types of data transformations and additional smoothing.
}
\examples{
create_multi_line_plot(data = monthly_data,
                       var_name_list = c("bank_assets_sa", "bank_loans", "bank_securities_sa"),
                       var_label_list = c("Bank Assets", "Bank Loans", "Bank Securities"),
                       start_date = "1980-01-01", end_date = "2023-12-31",
                       x_axis_breaks = 5, use_month_intervals = FALSE,
                       x_axis_title = "Year", y_axis_title = "Billions of Dollars",
                       title = "Commercial Banks", subtitle = "Financial Indicators Over Time",
                       var_changes = NULL, include_smooth = FALSE, lof_span = 0.1,
                       y_axis_lower_bound = NULL, y_axis_upper_bound = NULL,
                       recession_bars = FALSE, diagonal_x_labels = FALSE)
}
