% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter_plot.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Create a Scatter Plot}
\usage{
create_scatter_plot(
  data,
  x_var,
  y_var,
  x_label,
  y_label,
  title = NULL,
  subtitle = NULL,
  highlight_dates = NULL,
  highlight_latest = FALSE,
  include_lof = NULL,
  log_x = FALSE,
  log_y = FALSE,
  start_date = NULL,
  end_date = NULL,
  y_lag = NULL,
  x_change = NULL,
  y_change = NULL,
  date_ranges = NULL,
  lof_color = "black",
  dot_size = 1,
  highlight_size = 3
)
}
\arguments{
\item{data}{A data frame containing the dataset for plotting.}

\item{x_var}{Name of the variable to be plotted on the x-axis.}

\item{y_var}{Name of the variable to be plotted on the y-axis.}

\item{x_label}{Label for the x-axis.}

\item{y_label}{Label for the y-axis.}

\item{title}{Optional title for the plot.}

\item{subtitle}{Optional subtitle for the plot. Overrides any automatic subtitle generation.}

\item{highlight_dates}{Dates to highlight on the plot.}

\item{highlight_latest}{A boolean indicating whether to highlight the latest data point.}

\item{include_lof}{Method for line of best fit ('lm', 'loess', or NULL).}

\item{log_x}{A boolean indicating whether to apply a logarithmic transformation to the x-axis.}

\item{log_y}{A boolean indicating whether to apply a logarithmic transformation to the y-axis.}

\item{start_date}{The starting date for the plot (optional).}

\item{end_date}{The ending date for the plot (optional).}

\item{y_lag}{The number of rows to shift the y variable back or forward. If positive, the y-column is shifted up (behind) by y_lag.}

\item{x_change}{A string indicating whether to calculate a % change for the x variable: yoy, qoq, or mom. If NULL, then no change. Default NULL.}

\item{y_change}{A string indicating whether to calculate a % change for the y variable: yoy, qoq, or mom. If NULL, then no change. Default NULL.}

\item{date_ranges}{A list of date ranges to highlight on the plot.}

\item{lof_color}{Color for the line of best fit. This parameter allows you to specify the color of the line of best fit (LOF) on the scatter plot. The default color is "black". Accepts any color value recognized by ggplot2, such as color names ("red", "blue", etc.) or hexadecimal color codes.}

\item{dot_size}{Size of the normal (non-highlighted) dots on the scatter plot. This parameter controls the size of the standard points in the plot. The default size is 1. You can specify any positive numeric value to increase or decrease the size of these points.}

\item{highlight_size}{Size of the highlighted dots on the scatter plot. This parameter controls the size of the dots that are highlighted (e.g., through `highlight_dates`, `highlight_latest`, or `date_ranges`). The default size is 3. You can specify any positive numeric value to adjust the size of these highlighted points.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function generates a scatter plot for the relationship between two variables over time. It includes options for highlighting specific dates, adding a line of best fit, and log transformations.
}
\examples{
create_scatter_plot(data = monthly_data, x_var = "total_completed_nsa", y_var = "total_housing_starts_nsa",
                    x_label = "Completed (1000s)", y_label = "Starts (1000s)",
                    title = "Housings Starts vs Completed",
                    highlight_dates = c("2022-01-01", "2021-01-01", "2023-01-01", "2020-01-01", "2019-01-01", "2018-01-01"),
                    highlight_latest = TRUE,
                    include_lof = NULL,
                    log_x = FALSE, log_y = FALSE)
create_scatter_plot(data = monthly_data, x_var = "ngdp", y_var = "total_housing_starts_nsa",
                    x_label = "NGDP", y_label = "Starts",
                    title = "Housings Starts vs NGDP (Logs)",
                    highlight_dates = c("2022-01-01", "2021-01-01", "2023-01-01", "2020-01-01", "2019-01-01", "2018-01-01"),
                    highlight_latest = TRUE,
                    include_lof = "lm",
                    log_x = TRUE, log_y = TRUE)
}
