% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_two_axis_line_plot.R
\name{create_two_axis_line_plot}
\alias{create_two_axis_line_plot}
\title{Create a Two-Axis Line Plot}
\usage{
create_two_axis_line_plot(
  data,
  variables,
  var_labels,
  start_date,
  end_date,
  plot_title,
  x_axis_breaks,
  var_1_type = NULL,
  var_2_type = NULL,
  horizontal_line = "left"
)
}
\arguments{
\item{data}{A data frame containing the dataset for plotting.}

\item{variables}{A vector of two variable names in 'data' to be plotted.}

\item{var_labels}{A vector of two labels corresponding to the variables.}

\item{start_date}{The starting date for the plot, formatted as 'YYYY-MM-DD'.}

\item{end_date}{The ending date for the plot, formatted as 'YYYY-MM-DD'.}

\item{plot_title}{Title of the plot.}

\item{x_axis_breaks}{Breaks for the x-axis, usually years.}

\item{var_1_type}{Type of transformation for the first variable.
Acceptable values: 'yoy' for year-over-year, 'qoq' for quarter-over-quarter,
'mom' for month-over-month, or NULL for no transformation.}

\item{var_2_type}{Type of transformation for the second variable.
Acceptable values: 'yoy' for year-over-year, 'qoq' for quarter-over-quarter,
'mom' for month-over-month, or NULL for no transformation.}

\item{horizontal_line}{Defaulted to "left", which adds a horizontal line at the 0-line for the left y axis. Set to "right" to put the horizontal line at the 0-line for the right y axis.}
}
\value{
A ggplot object representing the two-axis line plot.
}
\description{
This function generates a line plot for two variables, each with its own y-axis, over a given time period. The plot includes recession bars and supports different types of data transformations.
}
\examples{
create_two_axis_line_plot(data = monthly_data, variables = c("fed_funds_rate", "ngdp_monthly"),
                          var_labels = c("Fed Funds Rate (\%)", "NGDP YoY \% Change"), start_date = "2000-01-01",
                          end_date = "2024-01-01", x_axis_breaks = 5, plot_title = "Fed Funds and NGDP",
                          var_1_type = NULL, var_2_type = "yoy")
}
